backend/requirements.txt
tensorflow==2.12.0
music21==7.3.3
numpy==1.23.5
pandas==1.5.3
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
python-dotenv==1.0.0
SQLAlchemy==1.4.41
flask_sqlalchemy==2.5.1